Code for star six.
Fixed a bug in the parse code: graph was rotated 90 degrees. This wouldn't matter for either distance metric, but it was a debug headache.
Added a magnitude tracker in each Line segment so it knows how far the wire has come up to its starting point.
Added logic to the intersection function to check if the segments are orthogonal. If not, the comparisons don't work correctly.
Returning the distance to the intersect point along with the actual point in the intersection function.